refactor: update fines permissions structure and relocate interfaces
