+++ src/app/flows/fines/fines-mac/fines-mac-payment-terms/interfaces/fines-mac-payment-terms-permissions.interface.ts
@@ -1,0 +1,3 @@
+export interface IFinesMacPaymentTermsPermissions {
+  [key: number]: boolean;
+}
