+++ src/app/interfaces/fines-permissions.interfaces.ts
@@ -1,6 +1,3 @@
 export interface IFinesPermissions {
   [key: string]: number;
 }
-export interface IFinesMacPaymentTermsPermissions {
-  [key: number]: boolean;
-}
